applying guards + redirect fix

+++ app/Http/Controllers/CommentController.php
@@ -56,6 +56,8 @@ class CommentController extends Controller
      */
     public function destroy(Comment $comment)
     {
+        $this->authorize('delete-comment' , $comment);
+        
         $comment->delete();
 
         return back();
